Terceiro Commit:  DeleteProduct e Busca de produtos pelo nome, preço máximo e mínimo feitas  Troquei o dadod do Preço de BigDecimal pra Double

+++ src/main/java/com/project/ecommerce/model/ProductModel.java
@@ -9,9 +9,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-
-import java.math.BigDecimal;
-
 @Entity
 @Data
 @AllArgsConstructor
@@ -24,7 +21,7 @@ public class ProductModel {
     private Long productId;
     private String productName;
     private String productDescription;
-    private BigDecimal productPrice;
+    private Double productPrice;
     private ProductClassification productClassification;
 
 }
